image/requirements.txt
requests==2.31.0
urllib3==1.26.6
pyyaml>=3.10
cfgv==3.3.1
terrasnek==0.1.13
pygithub==1.59.0
setuptools==68.0.0
